Center divs using two different methods
This exercise asks for a solution to center a div inside a container. The intended answer is to use justify-content and align-items, but one can also use "margin: auto" in the flex-item to center. If more than one flex-item exist, this solution will distribute them equally, which is also usefull!